Refactor catalogs into submodule
